backend/etl/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0